shiftright128's dist parameter shouldn't be 64-bit.
shiftright128() was defined as taking `uint64_t dist`, but this was
unnecessary. It's always called with `int32_t` or `uint32_t`, and it
requires `0 < dist < 128`.

Changing this parameter to `uint32_t dist` improves codegen
for 32-bit platforms (I observe MSVC x86 /O2 saving 25 bytes)
and doesn't affect codegen for 64-bit platforms.

+++ ryu/mulshift128.h
@@ -27,7 +27,7 @@ static inline uint64_t umul128(uint64_t a, uint64_t b, uint64_t* productHi) {
   return _umul128(a, b, productHi);
 }
 
-static inline uint64_t shiftright128(uint64_t lo, uint64_t hi, uint64_t dist) {
+static inline uint64_t shiftright128(uint64_t lo, uint64_t hi, uint32_t dist) {
   return __shiftright128(lo, hi, (unsigned char) dist);
 }
 
@@ -54,7 +54,7 @@ static inline uint64_t umul128(uint64_t a, uint64_t b, uint64_t* productHi) {
   return productLo;
 }
 
-static inline uint64_t shiftright128(uint64_t lo, uint64_t hi, uint64_t dist) {
+static inline uint64_t shiftright128(uint64_t lo, uint64_t hi, uint32_t dist) {
   // shift hi-lo right by 0 < dist < 128
   return (dist >= 64)
       ? hi >> (dist - 64)
